+++ packages/address/src/loader.ts
@@ -64,7 +64,7 @@ export const loadCountry: (
   },
 );
 
-class CountryLoaderError extends Error {
+export class CountryLoaderError extends Error {
   constructor(errors: ResponseError) {
     const errorMessage = errors.errors.map((error) => error.message).join('; ');
     super(errorMessage);