+++ packages/address-mocks/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- `mockCountryRequests` will now return the fixture for the country code passed in the variables if available or return a relevant GraphQL response error if not. [[#2179](https://github.com/Shopify/quilt/pull/2179)]
 
 ## 2.0.18 - 2022-03-09